Partial conversion from 0.0->1.0 texture coordinate range. For now only terrain textures are converted. Once they are done, the texture matrix hack is restored. Thanks to devurandom and cybersphinx for help and previous efforts in this area.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@2980 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_common/piedef.h
@@ -124,7 +124,7 @@ typedef union  {PIELIGHTBYTES byte; UDWORD argb;} PIELIGHT;
 typedef struct
 {
 	Vector3i pos;
-	unsigned int u, v;
+	float u, v;
 	PIELIGHT light, specular;
 	Vector3i screen; //! Screenspace tile coordinates
 	Vector3i water; //! Screenspace water coordinates
